@@ -1,11 +1,12 @@
-'use strict';
-const path = require('path');
-const BinWrapper = require('bin-wrapper');
-const pkg = require('../package.json');
+import fs from 'node:fs';
+import process from 'node:process';
+import {fileURLToPath} from 'node:url';
+import BinWrapper from 'bin-wrapper';
 
+const pkg = JSON.parse(fs.readFileSync(new URL('../package.json', import.meta.url)));
 const url = `https://raw.githubusercontent.com/imagemin/gifsicle-bin/v${pkg.version}/vendor/`;
 
-module.exports = new BinWrapper()
+const binWrapper = new BinWrapper()
 	.src(`${url}macos/gifsicle`, 'darwin')
 	.src(`${url}linux/x86/gifsicle`, 'linux', 'x86')
 	.src(`${url}linux/x64/gifsicle`, 'linux', 'x64')
@@ -13,5 +14,7 @@ module.exports = new BinWrapper()
 	.src(`${url}freebsd/x64/gifsicle`, 'freebsd', 'x64')
 	.src(`${url}win/x86/gifsicle.exe`, 'win32', 'x86')
 	.src(`${url}win/x64/gifsicle.exe`, 'win32', 'x64')
-	.dest(path.join(__dirname, '../vendor'))
+	.dest(fileURLToPath(new URL('../vendor', import.meta.url)))
 	.use(process.platform === 'win32' ? 'gifsicle.exe' : 'gifsicle');
+
+export default binWrapper;
