@@ -1,7 +1,7 @@
-'use strict';
-const path = require('path');
-const binBuild = require('bin-build');
-const bin = require('.');
+import process from 'node:process';
+import {fileURLToPath} from 'node:url';
+import binBuild from 'bin-build';
+import bin from './index.js';
 
 (async () => {
 	try {
@@ -14,14 +14,15 @@ const bin = require('.');
 
 		const config = [
 			'./configure --disable-gifview --disable-gifdiff',
-			`--prefix="${bin.dest()}" --bindir="${bin.dest()}"`
+			`--prefix="${bin.dest()}" --bindir="${bin.dest()}"`,
 		].join(' ');
 
 		try {
-			await binBuild.file(path.resolve(__dirname, '../vendor/source/gifsicle-1.92.tar.gz'), [
+			const source = fileURLToPath(new URL('../vendor/source/gifsicle-1.92.tar.gz', import.meta.url));
+			await binBuild.file(source, [
 				'autoreconf -ivf',
 				config,
-				'make install'
+				'make install',
 			]);
 
 			console.log('gifsicle built successfully');
