@@ -1,13 +1,14 @@
-'use strict';
-const fs = require('fs');
-const path = require('path');
-const test = require('ava');
-const execa = require('execa');
-const tempy = require('tempy');
-const binCheck = require('bin-check');
-const binBuild = require('bin-build');
-const compareSize = require('compare-size');
-const m = require('..');
+import fs from 'node:fs';
+import path from 'node:path';
+import process from 'node:process';
+import {fileURLToPath} from 'node:url';
+import test from 'ava';
+import execa from 'execa';
+import tempy from 'tempy';
+import binCheck from 'bin-check';
+import binBuild from 'bin-build';
+import compareSize from 'compare-size';
+import gifsicle from '../index.js';
 
 test.serial('rebuild the gifsicle binaries', async t => {
 	// Skip the test on Windows
@@ -19,33 +20,34 @@ test.serial('rebuild the gifsicle binaries', async t => {
 	const temporary = tempy.directory();
 	const cfg = [
 		'./configure --disable-gifview --disable-gifdiff',
-		`--prefix="${temporary}" --bindir="${temporary}"`
+		`--prefix="${temporary}" --bindir="${temporary}"`,
 	].join(' ');
 
-	await binBuild.file(path.resolve(__dirname, '../vendor/source/gifsicle-1.92.tar.gz'), [
+	const source = fileURLToPath(new URL('../vendor/source/gifsicle-1.92.tar.gz', import.meta.url));
+	await binBuild.file(source, [
 		'autoreconf -ivf',
 		cfg,
-		'make install'
+		'make install',
 	]);
 
 	t.true(fs.existsSync(path.join(temporary, 'gifsicle')));
 });
 
 test.serial('verify binary', async t => {
-	t.true(await binCheck(m, ['--version']));
+	t.true(await binCheck(gifsicle, ['--version']));
 });
 
 test.serial('minify a gif', async t => {
 	const temporary = tempy.directory();
-	const src = path.join(__dirname, 'fixtures/test.gif');
+	const src = fileURLToPath(new URL('fixtures/test.gif', import.meta.url));
 	const dest = path.join(temporary, 'test.gif');
 	const args = [
 		'-o',
 		dest,
-		src
+		src,
 	];
 
-	await execa(m, args);
+	await execa(gifsicle, args);
 	const result = await compareSize(src, dest);
 
 	t.true(result[dest] < result[src]);
